requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
boto3==1.37.5
botocore==1.37.5
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
DBUtils==3.1.0
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
mailjet-rest==1.3.4
MarkupSafe==3.0.2
pika==1.3.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.6.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.3
rococo==1.0.30
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
surrealdb==0.3.2
typing_extensions==4.12.2
urllib3==2.3.0
websockets==10.4
Werkzeug==3.1.3
